add core CMakeLists.tct file
This CmakeLists.txt does the heavy lifting. It keeps track of the
source dependencies and uses a python script to read in conversion
configurations. If the nvidia texture tools and ARM astcenc are
available it can transform from an input file format (e.g. PNG) to DXT1,
DXT3, DXT5 and ASTC.